pan/lib, panvk: Ensure data_size is on 64 bits
It was previously possible to overflow the data_size calculation (for
example with width=4096, height=4096, array_size=256 and any format)

This was then causing GetImageMemoryRequirements2 to return a size of 0.

Fix "dEQP-VK.pipeline.*.render_to_image.core.2d_array.huge.width_height_layers.r8g8b8a8_unorm" failures.

Acked-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/29670>

+++ src/panfrost/lib/pan_layout.c
@@ -470,8 +470,8 @@ pan_image_layout_init(unsigned arch, struct pan_image_layout *layout,
    if (explicit_layout)
       layout->data_size = offset;
    else
-      layout->data_size =
-         ALIGN_POT(layout->array_stride * layout->array_size, 4096);
+      layout->data_size = ALIGN_POT(
+         (uint64_t)layout->array_stride * (uint64_t)layout->array_size, 4096);
 
    return true;
 }
